+++ src/hmlr/components/header/header.stories.js
@@ -7,10 +7,6 @@ export default {
   argTypes: {
     serviceName: { control: "text" },
     serviceUrl: { control: "text" },
-    navigation: { control: "object" },
-    navigationClasses: { control: "text" },
-    navigationLabel: { control: "text" },
-    menuButtonLabel: { control: "text" },
     containerClasses: { control: "text" },
     classes: { control: "text" },
     attributes: { control: "text" },
@@ -20,10 +16,6 @@ export default {
 const Template = ({
   serviceName,
   serviceUrl,
-  navigation,
-  navigationClasses,
-  navigationLabel,
-  menuButtonLabel,
   containerClasses,
   classes,
   attributes,
@@ -32,10 +24,6 @@ const Template = ({
     params: {
       serviceName,
       serviceUrl,
-      navigation,
-      navigationClasses,
-      navigationLabel,
-      menuButtonLabel,
       containerClasses,
       classes,
       attributes,
@@ -48,31 +36,8 @@ Standard.args = {
   serviceName: "My Service",
 };
 
-// export const Navigation = Template.bind({});
-// Navigation.args = {
-//   serviceName: "My Service",
-//   navigation: [
-//     {
-//       text: "Item 1",
-//       href: "#",
-//       active: true,
-//       attributes: {
-//         foo: "bar",
-//       },
-//     },
-//     {
-//       text: "Item 2",
-//       href: "#",
-//       attributes: {
-//         foo: "bar",
-//       },
-//     },
-//     {
-//       html: "Item 3",
-//       href: "#",
-//       attributes: {
-//         foo: "bar",
-//       },
-//     },
-//   ],
-// };
+export const FullWidth = Template.bind({});
+FullWidth.args = {
+  serviceName: "My Service",
+  containerClasses: "govuk-header__container--full-width"
+};